Rename dir - 4 connection_detail
