+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/connection_detail/access_control_tab.tsx
