+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/connection_detail/utils/connection_management_callout.tsx
