+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/connection_detail/utils/inactive_data_connection_callout.test.tsx
