+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/connection_detail/utils/no_access_page.test.tsx
