+++ src/plugins/data_source_management/public/components/direct_query_data_sources_components/connection_detail/utils/save_or_cancel.tsx
