+++ src/plugins/data_source_management/public/management_app/mount_management_section.tsx
@@ -21,7 +21,7 @@ import { CreateDataSourcePanel } from '../components/data_source_creation_panel/
 import { DataSourceManagementContext } from '../types';
 import { AuthenticationMethodRegistry } from '../auth_registry';
 import { ConfigureDirectQueryDataSourceWithRouter } from '../components/direct_query_data_sources_components/direct_query_data_source_configuration/configure_direct_query_data_sources';
-import { DirectQueryDataConnectionDetail } from '../components/direct_query_data_sources_components/direct_query_connection_detail/direct_query_connection_detail';
+import { DirectQueryDataConnectionDetail } from '../components/direct_query_data_sources_components/connection_detail/direct_query_connection_detail';
 
 export interface DataSourceManagementStartDependencies {
   data: DataPublicPluginStart;
